+++ crates/affinidi-tdk/common/affinidi-did-authentication/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Affinidi DID Authentication
 
+## 21st April 2025 (0.1.7)
+
+* Improved logging of error when no auth service is found
+
 ## 24th March 2025 Release 0.1.6
 
 * Implemented caching expiry for AuthenticationRecords